move edge detection component to separate file
