+++ src/lib.rs
@@ -1,4 +1,5 @@
 mod camera;
+mod camera_edge_detection;
 mod camera_image_pipe;
 mod camera_node;
 mod camera_node_sobel;
@@ -8,11 +9,7 @@ mod widget;
 mod widget_halfblocks;
 mod widget_luminance;
 
+pub use camera::{LuminanceConfig, RatatuiCamera, RatatuiCameraStrategy};
+pub use camera_edge_detection::{EdgeCharacters, RatatuiCameraEdgeDetection};
 pub use plugin::RatatuiCameraPlugin;
-
-pub use camera::{
-    EdgeCharacters, LuminanceConfig, RatatuiCamera, RatatuiCameraEdgeDetection,
-    RatatuiCameraStrategy,
-};
-
 pub use widget::RatatuiCameraWidget;
